requirements.txt
psutil==5.9.4
python-nmap
pymongo
colorama
